+++ magento/types.py
@@ -37,11 +37,6 @@ MagentoEntity = JSONDict
 PathId = Union[int, str]
 Sku = str
 
-# TODO: proper types
-Category = MagentoEntity
-Customer = MagentoEntity
-Order = MagentoEntity
-
 
 class WithExtensionAttributesDict(TypedDict):
     """A dict with an extension_attributes key."""
@@ -123,6 +118,26 @@ class Product(WithExtensionAttributesDict):
     weight: NotRequired[float]
 
 
+# Categories
+# ==========
+
+class Category(WithExtensionAttributesDict):
+    """A category."""
+    id: int
+    parent_id: int
+    name: str
+    is_active: bool
+    position: int
+    level: int
+    children: str
+    created_at: str
+    updated_at: str
+    path: str
+    available_sort_by: List[str]
+    include_in_menu: bool
+    custom_attributes: List[CustomAttributeDict]
+
+
 # Source items
 # ============
 # See https://developer.adobe.com/commerce/webapi/rest/inventory/manage-source-items/
@@ -173,6 +188,40 @@ class ShippingAddress(TypedDict):
     country_id: str
     telephone: str
     company: NotRequired[str]
+
+
+# TODO: proper TypedDict
+Order = MagentoEntity
+
+
+# Customers
+# =========
+
+class Customer(WithExtensionAttributesDict):
+    """A customer."""
+    # Unclear what's really always present here
+    id: NotRequired[int]
+    group_id: NotRequired[int]
+    default_billing: NotRequired[str]
+    default_shipping: NotRequired[str]
+    confirmation: NotRequired[str]
+    created_at: str
+    updated_at: str
+    created_in: NotRequired[str]
+    dob: NotRequired[str]
+    email: str
+    firstname: str
+    lastname: str
+    middlename: NotRequired[str]
+    prefix: NotRequired[str]
+    suffix: NotRequired[str]
+    gender: NotRequired[int]
+    store_id: NotRequired[int]
+    taxvat: NotRequired[str]
+    website_id: NotRequired[int]
+    addresses: NotRequired[List[Any]]  # TODO typing
+    disable_auto_group_change: NotRequired[int]
+    custom_attributes: List[CustomAttributeDict]
 
 
 # Other types